+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/MultiplotRecipeBase.cs
@@ -1,34 +1,17 @@
-﻿using NUnit.Framework.Internal;
-
-namespace ScottPlotCookbook.Recipes;
+﻿namespace ScottPlotCookbook.Recipes;
 
 #pragma warning disable IDE1006 // Ignore lowercase public variable names
 
-/// <summary>
-/// Individual recipes can inherit this so they double as <see cref="Recipes.IRecipe"/> and
-/// test cases that have a function decorated with the <see cref="Test"/> attribute.
-/// </summary>
-public abstract class RecipeBase : IRecipe
+public abstract class MultiplotRecipeBase
 {
     public static int ImageWidth { get; set; } = 400;
 
-    public static int ImageHeight { get; set; } = 300;
-
+    public static int ImageHeight { get; set; } = 400;
     public abstract string Name { get; }
-
     public abstract string Description { get; }
 
     // keep this lowercase because it should be lowercase in recipe source code
-    public Plot myPlot { get; private set; } = new();
-
-    /// <summary>
-    /// This function is called by code interacting with <see cref="Recipes.IRecipe"/>
-    /// </summary>
-    public void Execute(Plot plot)
-    {
-        myPlot = plot;
-        Execute();
-    }
+    public Multiplot multiplot { get; private set; } = new();
 
     /// <summary>
     /// This function is called from within the test system
@@ -49,7 +32,7 @@ public abstract class RecipeBase : IRecipe
         string imageFilename = $"{recipeClassName}.png";
         string saveAs = Path.Combine(Paths.OutputImageFolder, imageFilename);
         Directory.CreateDirectory(Paths.OutputImageFolder);
-        myPlot.SavePng(saveAs, ImageWidth, ImageHeight);
+        multiplot.SavePng(saveAs, ImageWidth, ImageHeight);
         Console.WriteLine($"{saveAs}");
     }
 }