+++ src/ScottPlot5/ScottPlot5 Tests/Unit Tests/CodeTests/CsprojTests.cs
@@ -3,14 +3,35 @@
 namespace ScottPlotTests.CodeTests;
 internal class CsprojTests
 {
-    readonly static Dictionary<string, string[]> CsprojFiles = Directory
+    readonly static Dictionary<string, string[]> ProjectFileContents = Directory
         .GetFiles(Path.Join(Paths.RepoFolder, "src/ScottPlot5/"), "*.csproj", SearchOption.AllDirectories)
-        .ToDictionary(x => x, x => File.ReadAllLines(x));
+        .ToDictionary(x => x, File.ReadAllLines);
+
+    static Dictionary<string, string[]> NugetProjectFileContents => ProjectFileContents
+        .Where(x => IsNugetProject(x.Key))
+        .ToDictionary(x => x.Key, x => x.Value);
+
+    static bool IsNugetProject(string csprojFilePath)
+    {
+        string folder = Path.GetDirectoryName(csprojFilePath)!.Replace(Paths.RepoFolder, "").Replace("\\", "/");
+        if (folder == "/src/ScottPlot5/ScottPlot5")
+            return true;
+        if (folder.StartsWith("/src/ScottPlot5/ScottPlot5 Controls/"))
+            return true;
+        return false;
+    }
+
+    [Test]
+    public void Test_Csproj_FilesAreFound()
+    {
+        ProjectFileContents.Should().NotBeEmpty();
+        NugetProjectFileContents.Should().NotBeEmpty();
+    }
 
     [Test]
     public void Test_Csproj_NoFloatingVersions()
     {
-        foreach ((string file, string[] lines) in CsprojFiles)
+        foreach ((string file, string[] lines) in ProjectFileContents)
         {
             foreach (string line in lines)
             {
@@ -20,7 +41,7 @@ internal class CsprojTests
                 if (line.Contains("*"))
                 {
                     throw new InvalidDataException(
-                        "csproj files must not contain pacakge references with floating version numbers. \n" +
+                        "csproj files must not contain package references with floating version numbers. \n" +
                         $"Offending file: {file} \n" +
                         $"Offending line: {line.Trim()}");
                 }
@@ -33,7 +54,7 @@ internal class CsprojTests
     {
         Dictionary<string, string> versionsByFile = [];
 
-        foreach ((string file, string[] lines) in CsprojFiles)
+        foreach ((string file, string[] lines) in ProjectFileContents)
         {
             foreach (string line in lines)
             {
@@ -62,11 +83,47 @@ internal class CsprojTests
     [Test]
     public void Test_Projects_AreSigned()
     {
-        foreach ((string file, string[] lines) in CsprojFiles)
+        foreach ((string file, string[] lines) in ProjectFileContents)
         {
             if (!lines.Where(x => x.Contains("<SignAssembly>")).Any())
                 Assert.Fail($"{file} requires 'SignAssembly' to be defined. " +
                     $"See other project files in this repository for more information.");
+        }
+    }
+
+    private static bool Contains(string search, string[] lines)
+    {
+        foreach (string line in lines)
+        {
+            if (line.Contains(search))
+                return true;
+        }
+        return false;
+    }
+
+    [Test]
+    public void Test_ProjectFiles_HaveMitLicenseExpression()
+    {
+        // https://learn.microsoft.com/en-us/nuget/create-packages/package-authoring-best-practices#licensing
+        foreach ((string file, string[] lines) in NugetProjectFileContents)
+        {
+            if (!Contains("<PackageLicenseExpression>MIT</PackageLicenseExpression>", lines))
+            {
+                Assert.Fail($"MIT license expression must be in all NuGet csproj files.\n{file}");
+            }
+        }
+    }
+
+    [Test]
+    public void Test_ProjectFiles_HaveCopyright()
+    {
+        // https://learn.microsoft.com/en-us/nuget/create-packages/package-authoring-best-practices#copyright
+        foreach ((string file, string[] lines) in NugetProjectFileContents)
+        {
+            if (!Contains("<Copyright>Copyright (c) Scott Harden / Harden Technologies, LLC</Copyright>", lines))
+            {
+                Assert.Fail($"Copyright must be in all NuGet csproj files.\n{file}");
+            }
         }
     }
 }