+++ src/ScottPlot5/ScottPlot5 Tests/Unit Tests/UnitTests/TickGenerators/DecimalTickSpacingCalculatorTests.cs
@@ -18,4 +18,18 @@ internal class DecimalTickSpacingCalculatorTests
             spacePerLabel.Should().BeGreaterThan(maxLabelLength.Length);
         }
     }
+
+    [Test]
+    public void GenerateTickPositions__SmallAxisSize_2DifferentTickPostitions()
+    {
+        ScottPlot.TickGenerators.DecimalTickSpacingCalculator calc = new();
+        CoordinateRange range = new(1, 5);
+        PixelLength axisLength = new(50);
+        PixelLength maxLabelLength = 30;
+
+        double[] positions = calc.GenerateTickPositions(range, axisLength, maxLabelLength);
+
+        Assert.That(positions.Length, Is.EqualTo(2));
+        Assert.That(positions[0], Is.Not.EqualTo(positions[1]));
+    }
 }